requirements.txt
Flask==3.1.0
minio==7.2.12
python-dotenv==1.0.1
